Fixed an issue in isFaceing
Casting it to int8 before absing lets us get the absolute
difference in the angles properly (before, values over 128 would
not start counting as closer [as they would since they are closer to
256])

Fixes #3891

+++ src/common/utils.cpp
@@ -187,7 +187,7 @@ uint8 getangle(const position_t& A, const position_t& B)
 bool isFaceing(const position_t& A, const position_t& B, uint8 coneAngle)
 {
 	int32 angle = getangle(A,B);
-	return ( abs(angle - A.rotation) < (coneAngle >> 1) );
+    return abs(int8(angle - A.rotation)) < (coneAngle >> 1);
 }
 
 /**
